Fix debug output broken in eed35e2.

+++ src/view_filebrowser.cc
@@ -909,7 +909,8 @@ void ViewFileBrowser::View::serialize(DCP::Queue &queue, DCP::Queue::Mode mode,
             *debug_os << "    ";
 
         if(item != nullptr)
-            *debug_os << "Type " << item->get_kind().get_raw_code() << " " << it << ": "
+            *debug_os << "Type " << (unsigned int)item->get_kind().get_raw_code()
+                      << " " << it << ": "
                       << item->get_text() << std::endl;
         else
             *debug_os << "*NULL ENTRY* " << it << std::endl;
